backend/generate_ssl_cert.py
"""
Generate self-signed SSL certificate for local HTTPS development
Run this with: python backend/generate_ssl_cert.py
"""
from cryptography import x509
from cryptography.x509.oid import NameOID
from cryptography.hazmat.primitives import hashes
from cryptography.hazmat.primitives.asymmetric import rsa
from cryptography.hazmat.primitives import serialization
import datetime
import os

# Generate private key
private_key = rsa.generate_private_key(
    public_exponent=65537,
    key_size=2048,
)

# Generate certificate
subject = issuer = x509.Name([
    x509.NameAttribute(NameOID.COUNTRY_NAME, "US"),
    x509.NameAttribute(NameOID.STATE_OR_PROVINCE_NAME, "Local"),
    x509.NameAttribute(NameOID.LOCALITY_NAME, "Local"),
    x509.NameAttribute(NameOID.ORGANIZATION_NAME, "LifeOS"),
    x509.NameAttribute(NameOID.COMMON_NAME, "localhost"),
])

cert = x509.CertificateBuilder().subject_name(
    subject
).issuer_name(
    issuer
).public_key(
    private_key.public_key()
).serial_number(
    x509.random_serial_number()
).not_valid_before(
    datetime.datetime.utcnow()
).not_valid_after(
    datetime.datetime.utcnow() + datetime.timedelta(days=365)
).add_extension(
    x509.SubjectAlternativeName([
        x509.DNSName("localhost"),
        x509.DNSName("127.0.0.1"),
    ]),
    critical=False,
).sign(private_key, hashes.SHA256())

# Write certificate to file
current_dir = os.path.dirname(os.path.abspath(__file__))
cert_path = os.path.join(current_dir, "cert.pem")
key_path = os.path.join(current_dir, "key.pem")

with open(cert_path, "wb") as f:
    f.write(cert.public_bytes(serialization.Encoding.PEM))

with open(key_path, "wb") as f:
    f.write(private_key.private_bytes(
        encoding=serialization.Encoding.PEM,
        format=serialization.PrivateFormat.TraditionalOpenSSL,
        encryption_algorithm=serialization.NoEncryption()
    ))

print("✓ SSL certificate generated successfully!")
print(f"  Certificate: {cert_path}")
print(f"  Private key: {key_path}")
