requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
httpx==0.28.1
python-dotenv==1.0.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
garminconnect==0.2.8
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.156.0
